ADD: Upon trembling ground we stand, to behold the fathomless depths of thy world.

+++ nn/inc/ops/softmax.h
@@ -11,12 +11,12 @@
 static inline void NN__softmax_F32(size_t n, float *y, float *x, size_t stride) {
   // exp and sum
   float sum = 0.0f;
-  for (int i = 0; i < n * stride; i += stride) {
+  for (size_t i = 0; i < n * stride; i += stride) {
     y[i] = expf(x[i]);
     sum += y[i];
   }
   // normalize
-  for (int i = 0; i < n * stride; i += stride) {
+  for (size_t i = 0; i < n * stride; i += stride) {
     y[i] /= sum;
   }
 }
